C++: Split off stl.h from stl.cpp.
